24 - 5

+++ 02 OOP/24-05 006 LTHDT Bai 3/Triangle.cs
@@ -50,9 +50,10 @@ namespace AAA
         {
             if (IsTriangle())
             {
-                double p = ((double)a + b + c) / 2;
-                return Math.Sqrt(p * (p - a) * (p - b) * (p - c)); }
-                
+                double p = ((double)a + (double)b + (double)c) / 2;
+                return Math.Sqrt(p * (p - a) * (p - b) * (p - c));
+            }
+
             else
             {
                 return 0; }
